Add DeletedChildsPlaceholder to the bottom of the comments container
